rename files to reduce the diff size
