+++ connection.go
@@ -6,12 +6,11 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at http://mozilla.org/MPL/2.0/.
 
-// +build go1.8
+// +build !go1.8
 
 package mysql
 
 import (
-	"context"
 	"database/sql/driver"
 	"net"
 	"strconv"
@@ -44,7 +43,7 @@ func (mc *mysqlConn) handleParams() (err error) {
 			charsets := strings.Split(val, ",")
 			for i := range charsets {
 				// ignore errors here - a charset may not exist
-				err = mc.exec(context.Background(), "SET NAMES "+charsets[i])
+				err = mc.exec("SET NAMES " + charsets[i])
 				if err == nil {
 					break
 				}
@@ -55,7 +54,7 @@ func (mc *mysqlConn) handleParams() (err error) {
 
 		// System Vars
 		default:
-			err = mc.exec(context.Background(), "SET "+param+"="+val+"")
+			err = mc.exec("SET " + param + "=" + val + "")
 			if err != nil {
 				return
 			}
@@ -65,36 +64,12 @@ func (mc *mysqlConn) handleParams() (err error) {
 	return
 }
 
-// Begin implements driver.Conn interface
 func (mc *mysqlConn) Begin() (driver.Tx, error) {
-	return mc.ConnBeginTx(context.Background(), driver.TxOptions{})
-}
-
-// Ping implements drvier.Pinger interface
-func (mc *mysqlConn) Ping(ctx context.Context) error {
-	if mc.netConn == nil {
-		errLog.Print(ErrInvalidConn)
-		return driver.ErrBadConn
-	}
-	if err := mc.writeCommandPacket(ctx, comPing); err != nil {
-		errLog.Print(err)
-		return err
-	}
-
-	if _, err := mc.readResultOK(); err != nil {
-		errLog.Print(err)
-		return err
-	}
-	return nil
-}
-
-// ConnBeginTx implements driver.ConnBeginTx interface
-func (mc *mysqlConn) ConnBeginTx(ctx context.Context, opts driver.TxOptions) (driver.Tx, error) {
 	if mc.netConn == nil {
 		errLog.Print(ErrInvalidConn)
 		return nil, driver.ErrBadConn
 	}
-	err := mc.exec(ctx, "START TRANSACTION")
+	err := mc.exec("START TRANSACTION")
 	if err == nil {
 		return &mysqlTx{mc}, err
 	}
@@ -105,7 +80,7 @@ func (mc *mysqlConn) ConnBeginTx(ctx context.Context, opts driver.TxOptions) (dr
 func (mc *mysqlConn) Close() (err error) {
 	// Makes Close idempotent
 	if mc.netConn != nil {
-		err = mc.writeCommandPacket(context.Background(), comQuit)
+		err = mc.writeCommandPacket(comQuit)
 	}
 
 	mc.cleanup()
@@ -130,16 +105,12 @@ func (mc *mysqlConn) cleanup() {
 }
 
 func (mc *mysqlConn) Prepare(query string) (driver.Stmt, error) {
-	return mc.PrepareContext(context.Background(), query)
-}
-
-func (mc *mysqlConn) PrepareContext(ctx context.Context, query string) (driver.Stmt, error) {
 	if mc.netConn == nil {
 		errLog.Print(ErrInvalidConn)
 		return nil, driver.ErrBadConn
 	}
 	// Send command
-	err := mc.writeCommandPacketStr(ctx, comStmtPrepare, query)
+	err := mc.writeCommandPacketStr(comStmtPrepare, query)
 	if err != nil {
 		return nil, err
 	}
@@ -288,10 +259,6 @@ func (mc *mysqlConn) interpolateParams(query string, args []driver.Value) (strin
 }
 
 func (mc *mysqlConn) Exec(query string, args []driver.Value) (driver.Result, error) {
-	return mc.ExecContext(context.Background(), query, args)
-}
-
-func (mc *mysqlConn) ExecContext(ctx context.Context, query string, args []driver.Value) (driver.Result, error) {
 	if mc.netConn == nil {
 		errLog.Print(ErrInvalidConn)
 		return nil, driver.ErrBadConn
@@ -311,7 +278,7 @@ func (mc *mysqlConn) ExecContext(ctx context.Context, query string, args []drive
 	mc.affectedRows = 0
 	mc.insertId = 0
 
-	err := mc.exec(ctx, query)
+	err := mc.exec(query)
 	if err == nil {
 		return &mysqlResult{
 			affectedRows: int64(mc.affectedRows),
@@ -322,9 +289,9 @@ func (mc *mysqlConn) ExecContext(ctx context.Context, query string, args []drive
 }
 
 // Internal function to execute commands
-func (mc *mysqlConn) exec(ctx context.Context, query string) error {
+func (mc *mysqlConn) exec(query string) error {
 	// Send command
-	err := mc.writeCommandPacketStr(ctx, comQuery, query)
+	err := mc.writeCommandPacketStr(comQuery, query)
 	if err != nil {
 		return err
 	}
@@ -342,13 +309,7 @@ func (mc *mysqlConn) exec(ctx context.Context, query string) error {
 	return err
 }
 
-// Query implements driver.Queryer interface
 func (mc *mysqlConn) Query(query string, args []driver.Value) (driver.Rows, error) {
-	return mc.QueryContext(context.Background(), query, args)
-}
-
-// QueryContext implements driver.QueryerContext interface
-func (mc *mysqlConn) QueryContext(ctx context.Context, query string, args []driver.Value) (driver.Rows, error) {
 	if mc.netConn == nil {
 		errLog.Print(ErrInvalidConn)
 		return nil, driver.ErrBadConn
@@ -366,7 +327,7 @@ func (mc *mysqlConn) QueryContext(ctx context.Context, query string, args []driv
 		args = nil
 	}
 	// Send command
-	err := mc.writeCommandPacketStr(ctx, comQuery, query)
+	err := mc.writeCommandPacketStr(comQuery, query)
 	if err == nil {
 		// Read Result
 		var resLen int
@@ -391,7 +352,7 @@ func (mc *mysqlConn) QueryContext(ctx context.Context, query string, args []driv
 // The returned byte slice is only valid until the next read
 func (mc *mysqlConn) getSystemVar(name string) ([]byte, error) {
 	// Send command
-	if err := mc.writeCommandPacketStr(context.Background(), comQuery, "SELECT @@"+name); err != nil {
+	if err := mc.writeCommandPacketStr(comQuery, "SELECT @@"+name); err != nil {
 		return nil, err
 	}
