+++ statement.go
@@ -1,5 +1,3 @@
-// +build go1.8
-
 // Go MySQL Driver - A MySQL-Driver for Go's database/sql package
 //
 // Copyright 2012 The Go-MySQL-Driver Authors. All rights reserved.
@@ -8,10 +6,11 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at http://mozilla.org/MPL/2.0/.
 
+// +build !go1.8
+
 package mysql
 
 import (
-	"context"
 	"database/sql/driver"
 	"fmt"
 	"reflect"
@@ -25,7 +24,6 @@ type mysqlStmt struct {
 	columns    []mysqlField // cached from the first query
 }
 
-// Close implements driver.Conn interface
 func (stmt *mysqlStmt) Close() error {
 	if stmt.mc == nil || stmt.mc.netConn == nil {
 		// driver.Stmt.Close can be called more than once, thus this function
@@ -35,12 +33,11 @@ func (stmt *mysqlStmt) Close() error {
 		return driver.ErrBadConn
 	}
 
-	err := stmt.mc.writeCommandPacketUint32(context.Background(), comStmtClose, stmt.id)
+	err := stmt.mc.writeCommandPacketUint32(comStmtClose, stmt.id)
 	stmt.mc = nil
 	return err
 }
 
-// NumInput implements driver.Stmt interface
 func (stmt *mysqlStmt) NumInput() int {
 	return stmt.paramCount
 }
@@ -49,20 +46,13 @@ func (stmt *mysqlStmt) ColumnConverter(idx int) driver.ValueConverter {
 	return converter{}
 }
 
-// Exec implements driver.Stmt interface
 func (stmt *mysqlStmt) Exec(args []driver.Value) (driver.Result, error) {
-	return stmt.ExecContext(context.Background(), args)
-}
-
-// ExecContent implements driver.StmtExecContext interface
-func (stmt *mysqlStmt) ExecContext(ctx context.Context, args []driver.Value) (driver.Result, error) {
-
 	if stmt.mc.netConn == nil {
 		errLog.Print(ErrInvalidConn)
 		return nil, driver.ErrBadConn
 	}
 	// Send command
-	err := stmt.writeExecutePacket(ctx, args)
+	err := stmt.writeExecutePacket(args)
 	if err != nil {
 		return nil, err
 	}
@@ -96,19 +86,13 @@ func (stmt *mysqlStmt) ExecContext(ctx context.Context, args []driver.Value) (dr
 	return nil, err
 }
 
-// Query implements driver.Stmt interface
 func (stmt *mysqlStmt) Query(args []driver.Value) (driver.Rows, error) {
-	return stmt.QueryContext(context.Background(), args)
-}
-
-// QueryContext implements driver.StmtQueryContext interface
-func (stmt *mysqlStmt) QueryContext(ctx context.Context, args []driver.Value) (driver.Rows, error) {
 	if stmt.mc.netConn == nil {
 		errLog.Print(ErrInvalidConn)
 		return nil, driver.ErrBadConn
 	}
 	// Send command
-	err := stmt.writeExecutePacket(ctx, args)
+	err := stmt.writeExecutePacket(args)
 	if err != nil {
 		return nil, err
 	}
